Split building_preview_mesh.rs (746 lines) into subdirectory modules
Replace the single 746-line file with a building_preview_mesh/ directory
containing four focused modules:

- mod.rs (108 lines): resource definition, startup system, plugin, re-exports
- mesh_data.rs (198 lines): PreviewMeshData builder, cuboid/roof primitives,
  lighten/darken color helpers
- generators.rs (252 lines): per-zone-type procedural mesh generators
- tests.rs (215 lines): all unit tests

All public items (BuildingPreviewMeshes, setup_building_preview_meshes,
BuildingPreviewMeshPlugin) remain re-exported from the module root so
external call sites are unchanged.

Part of #1389
